+++ analytics/pubmatic/logger.go
@@ -1,4 +1,4 @@
-package openwrap
+package pubmatic
 
 import (
 	"encoding/json"
